fix more warnings

+++ src/helpers.c
@@ -22,7 +22,7 @@ ncm_helpers_show_song_time(int32 length, char *buffer, int32 buffer_cap) {
     return ncm_song_show_time(length, buffer, buffer_cap);
 }
 
-static int64
+static int32
 menu_item_count(NcMenu *menu, enum NcMenuItemSource source) {
     if (menu == NULL) {
         return 0;
@@ -41,7 +41,7 @@ menu_item_count(NcMenu *menu, enum NcMenuItemSource source) {
 
 static bool
 menu_position_is_selected(NcMenu *menu, enum NcMenuItemSource source,
-                          int64 pos) {
+                          int32 pos) {
     uint32 flags;
     void *item;
 
@@ -65,7 +65,7 @@ menu_position_is_selected(NcMenu *menu, enum NcMenuItemSource source,
 
 static bool
 menu_set_position_selected(NcMenu *menu, enum NcMenuItemSource source,
-                           int64 pos, bool selected) {
+                           int32 pos, bool selected) {
     uint32 flags;
     void *item;
 
@@ -92,11 +92,11 @@ menu_set_position_selected(NcMenu *menu, enum NcMenuItemSource source,
 
 void
 ncm_menu_reverse_selection(NcMenu *menu, enum NcMenuItemSource source) {
-    int64 count;
+    int32 count;
     bool selected;
 
     count = menu_item_count(menu, source);
-    for (int64 i = 0; i < count; i += 1) {
+    for (int32 i = 0; i < count; i += 1) {
         selected = menu_position_is_selected(menu, source, i);
         menu_set_position_selected(menu, source, i, !selected);
     }
@@ -105,9 +105,9 @@ ncm_menu_reverse_selection(NcMenu *menu, enum NcMenuItemSource source) {
 
 bool
 ncm_menu_find_selected_range(NcMenu *menu, enum NcMenuItemSource source,
-                             int64 *first, int64 *last) {
-    int64 count;
-    int64 range_first;
+                             int32 *first, int32 *last) {
+    int32 count;
+    int32 range_first;
 
     count = menu_item_count(menu, source);
     if (first) {
@@ -118,7 +118,7 @@ ncm_menu_find_selected_range(NcMenu *menu, enum NcMenuItemSource source,
     }
 
     range_first = count;
-    for (int64 i = 0; i < count; i += 1) {
+    for (int32 i = 0; i < count; i += 1) {
         if (menu_position_is_selected(menu, source, i)) {
             range_first = i;
             break;
@@ -131,7 +131,7 @@ ncm_menu_find_selected_range(NcMenu *menu, enum NcMenuItemSource source,
     if (first) {
         *first = range_first;
     }
-    for (int64 i = count - 1; i >= range_first; i -= 1) {
+    for (int32 i = count - 1; i >= range_first; i -= 1) {
         if (menu_position_is_selected(menu, source, i)) {
             if (last) {
                 *last = i + 1;
@@ -144,9 +144,9 @@ ncm_menu_find_selected_range(NcMenu *menu, enum NcMenuItemSource source,
 
 bool
 ncm_menu_find_full_selected_range(NcMenu *menu, enum NcMenuItemSource source,
-                                  int64 *first, int64 *last) {
-    int64 range_first;
-    int64 range_last;
+                                  int32 *first, int32 *last) {
+    int32 range_first;
+    int32 range_last;
 
     if (!ncm_menu_find_selected_range(menu, source, &range_first,
                                       &range_last)) {
@@ -158,7 +158,7 @@ ncm_menu_find_full_selected_range(NcMenu *menu, enum NcMenuItemSource source,
         }
         return true;
     }
-    for (int64 i = range_first; i < range_last; i += 1) {
+    for (int32 i = range_first; i < range_last; i += 1) {
         if (!menu_position_is_selected(menu, source, i)) {
             return false;
         }
